[IMP] accounting/l10n_br: Resourceful links section for onboarding
Add the relevant links to improve the onboarding of new customers. This documentation is accessible from the `res.config` in the customer DB.

To improve the onboarding, when the customer clicks in the DB is redirected here, and the first thing they see are the videos and the onboarding document with more details on how to configure the Avalara Portal, Use cases, tips, and tricks.